+++ DjangoLibrary/__init__.py
@@ -306,3 +306,27 @@ user.save()""".format(
                 msg = msg + '\n\n' + traceback
             raise requests.exceptions.HTTPError(msg, response=response)
         return response.raise_for_status()
+
+    def query(self, model, **kwargs):
+        """Query the Django ORM.
+        """
+        url = 'http://{}:{}'.format(
+            self.host,
+            self.port
+        )
+        payload = {
+            'MODEL_PATH': model,
+            'QUERY_ARGS': json.dumps(kwargs)
+        }
+        response = requests.get(url, params=payload)
+        if response.status_code == 200:
+            return response.json()
+        status_code_400 = response.status_code == 400
+        type_json = response.headers.get('Content-Type') == 'application/json'
+        if status_code_400 and type_json:
+            msg = response.json().get('error', '')
+            traceback = response.json().get('traceback', '')
+            if traceback:
+                msg = msg + '\n\n' + traceback
+            raise requests.exceptions.HTTPError(msg, response=response)
+        return response.raise_for_status()